fixed create, update, get, delete movie

+++ src/main/java/com/pw/cinema/movie/MovieService.java
@@ -1,83 +1,78 @@
 package com.pw.cinema.movie;
 
-import com.pw.cinema.exceptions.AlreadyExistsException;
 import com.pw.cinema.movie_category.MovieCategory;
 import com.pw.cinema.movie_category.MovieCategoryRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
+import java.util.stream.Collectors;
+
+import static com.pw.cinema.Utils.response;
 
 @Service
 public class MovieService {
-
     private final MovieRepository movieRepository;
 
-    public MovieService(MovieRepository movieRepository) {
+    private final MovieCategoryRepository movieCategoryRepository;
+
+    public MovieService(MovieRepository movieRepository, MovieCategoryRepository movieCategoryRepository) {
         this.movieRepository = movieRepository;
+        this.movieCategoryRepository = movieCategoryRepository;
     }
 
 
-    public Object create(Movie movie) throws AlreadyExistsException {
-        if (movieRepository.findByTitle(movie.getTitle()) != null) {
-            throw new AlreadyExistsException("Movie with that name already exists");
-        }
-        Map<String, Object> response = new HashMap<>();
-        response.put("data", movieRepository.save(movie));
-        response.put("message", "Successfully added new movie");
-        response.put("success", true);
-        return response;
+    public Object create(Movie movie) {
+        Set<Long> movieCategoryIds =
+                movie.getCategory().stream().map(MovieCategory::getId).collect(Collectors.toSet());
+        if (!movieCategoryRepository.existsAllByIdIn(movieCategoryIds))
+            throw new NoSuchElementException("Not found categories.");
+        movie.setCategory(movieCategoryRepository.findAllByIdIn(movieCategoryIds));
+        Movie newMovie = movieRepository.save(movie);
+        return response(newMovie, "Successfully added new movie");
     }
 
     public Object getMovieList() {
-        List<Movie> movies = new ArrayList<>(movieRepository.findAll());
-        Map<String, Object> response = new HashMap<>();
-        response.put("data", movies);
-        response.put("message", "Successfully found movies");
-        response.put("success", true);
-        return response;
+        List<Movie> movies = movieRepository.findAll();
+        return response(movies, "Successfully found movies");
     }
 
-    public Object updateMovie(Movie movieChanges, Long id) {
-        Map<String, Object> response = new HashMap<>();
-        response.put("data", movieRepository.save(movieChanges));
-        response.put("message", "Successfully found movies");
-        response.put("success", true);
-        return response;
+    public Object updateMovie(Movie updatedMovie, Long id) {
+        if (!movieRepository.existsById(id))
+            throw new NoSuchElementException("Not found movie with this id.");
+        Set<Long> movieCategoryIds =
+                updatedMovie.getCategory().stream().map(MovieCategory::getId).collect(Collectors.toSet());
+        if (!movieCategoryRepository.existsAllByIdIn(movieCategoryIds))
+            throw new NoSuchElementException("Not found categories.");
+        updatedMovie.setId(id);
+        updatedMovie.setCategory(movieCategoryRepository.findAllByIdIn(movieCategoryIds));
+        Movie savedMovie = movieRepository.save(updatedMovie);
+        return response(savedMovie, "Successfully found movies");
     }
 
     public Object deleteMovie(Long id) {
-        Map<String, Object> response = new HashMap<>();
+        if (!movieRepository.existsById(id))
+            throw new NoSuchElementException("Not found movie with this id.");
         movieRepository.deleteById(id);
+        Map<String, Object> response = new HashMap<>();
         response.put("message", "Successfully deleted movie");
         response.put("success", true);
         return response;
     }
 
     public Object getMovieById(Long id) {
+        Movie movie = movieRepository.findById(id).orElseThrow(() -> new NoSuchElementException("No movie present with " +
+                "id " + id));
+        return response(movie, "Successfully found movie");
+    }
+
+    public Object uploadPoster(Long id, String poster) {
+        //TODO
         Movie movie = movieRepository.findById(id).get();
         Map<String, Object> response = new HashMap<>();
-        response.put("data", movie);
+        movie.setPosterPhoto(poster);
+        response.put("data", movieRepository.save(movie));
         response.put("message", "Successfully found movie");
         response.put("success", true);
         return response;
     }
-
-//    public Object uploadPoster(String poster, Long id) {
-//        Movie movie = movieRepository.findById(id).get();
-//        Map<String, Object> response = new HashMap<>();
-//        movie.setPosterPhoto(poster);
-//        response.put("data", movieRepository.save(movie));
-//        response.put("message", "Successfully found movie");
-//        response.put("success", true);
-//        return response;
-//    }
-//
-//    public Object addCategoryToMovie(Long movieId, Long categoryId) {
-//        Movie movie = movieRepository.findById(movieId).get();
-//        MovieCategory movieCategory = movieCategoryRepository.findById(categoryId).get();
-//        movie.addCategory(movieCategory);
-//        return movieRepository.save(movie);
-//    }
 }
